+++ catalog/app/containers/Bucket/Dir/Toolbar/CreatePackage/useSuccessors.ts
@@ -3,7 +3,6 @@ import * as React from 'react'
 import { workflowsConfig } from 'containers/Bucket/requests'
 import * as AWS from 'utils/AWS'
 import * as Request from 'utils/useRequest'
-import { bucketToSuccessor } from 'utils/workflows'
 import type { Successor } from 'utils/workflows'
 
 export default function useSuccessors(bucket: string): Request.Result<Successor[]> {
@@ -15,7 +14,5 @@ export default function useSuccessors(bucket: string): Request.Result<Successor[
     return data
   }
 
-  // If successors are defined in config, use them as-is (don't auto-add current bucket)
-  // If no successors defined, add current bucket as default
-  return data.successors.length > 0 ? data.successors : [bucketToSuccessor(bucket)]
+  return data.successors
 }