+++ catalog/app/containers/Bucket/requests/object.ts
@@ -207,18 +207,20 @@ export const WORKFLOWS_CONFIG_PATH = quiltConfigs.workflows
 interface WorkflowsConfigArgs {
   s3: S3
   bucket: string
+  strict?: boolean
 }
 
-export const workflowsConfig = async ({ s3, bucket }: WorkflowsConfigArgs) => {
+export const workflowsConfig = async ({ s3, bucket, strict }: WorkflowsConfigArgs) => {
   try {
     const response = await fetchFile({
       s3,
       handle: { bucket, key: WORKFLOWS_CONFIG_PATH },
     })
-    return workflows.parse(response.body?.toString('utf-8') || '')
+    return workflows.parse(response.body?.toString('utf-8') || '', bucket, { strict })
   } catch (e) {
-    if (e instanceof FileNotFound || e instanceof VersionNotFound)
-      return workflows.emptyConfig
+    if (e instanceof FileNotFound || e instanceof VersionNotFound) {
+      return strict ? workflows.nullConfig : workflows.emptyConfig(bucket)
+    }
 
     Log.info('Unable to fetch')
     Log.error(e)